Optimal moves table for 8-puzzle
A lookup table maps each board position to the minimum number of moves
required to solve that particular position.

The table is generated via a breadth-first search. Root of the search
tree is the solved position with zero as the number of moves.

Between each iteration of the search (all boards 2 moves away from,
all boards 3 moves away, etc) control is yielded back to the UI
thread via a setTimeout() call.